scene_graph_window.cpp
